Rename Alaram Clock to Alarm Clock
This code uses the os, datetime, tkinter, time, and sys modules in Python to create the alarm clock functionality. It sets up a function get_time() for displaying the digital clock and function setting_alarm_time() that continuously checks the current time and triggers the alarm when the specified alarm time is reached.

The while loop compares the current time with the alarm time. If they match, it prints a message (Time to wake up!") and plays a sound using the os.startfile() function. You can replace "D:\grow with code\Task1\Twirling-alarm-tone.mp3" with the file path of your preferred alarm sound.

In the last part of the code, the get_time() function is called for displaying the current digital clock.